Simplify DustAppRun action
