+++ front/components/actions/dust_app_run/DustAppRunActionDetails.tsx
@@ -17,8 +17,8 @@ import { useMemo } from "react";
 import { ActionDetailsWrapper } from "@app/components/actions/ActionDetailsWrapper";
 import type { ActionDetailsComponentBaseProps } from "@app/components/actions/types";
 import { DUST_CONVERSATION_HISTORY_MAGIC_INPUT_KEY } from "@app/lib/actions/constants";
-import type { DustAppRunActionType } from "@app/lib/actions/types/dust_app_run";
-import { getDustAppRunResultsFileTitle } from "@app/lib/actions/types/dust_app_run";
+import type { DustAppRunActionType } from "@app/lib/actions/dust_app_run";
+import { getDustAppRunResultsFileTitle } from "@app/lib/actions/dust_app_run";
 import type { SupportedFileContentType } from "@app/types";
 
 function ContentTypeIcon({
