+++ front/components/assistant_builder/server_side_props_helpers.ts
@@ -9,9 +9,9 @@ import {
   getDefaultWebsearchActionConfiguration,
 } from "@app/components/assistant_builder/types";
 import { REASONING_MODEL_CONFIGS } from "@app/components/providers/types";
+import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run";
 import type { TablesQueryConfigurationType } from "@app/lib/actions/tables_query";
 import type { AgentActionConfigurationType } from "@app/lib/actions/types/agent";
-import type { DustAppRunConfigurationType } from "@app/lib/actions/types/dust_app_run";
 import {
   isBrowseConfiguration,
   isDustAppRunConfiguration,
