+++ front/lib/actions/configuration/dust_app_run.ts
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import { Op } from "sequelize";
 
-import type { DustAppRunConfigurationType } from "@app/lib/actions/types/dust_app_run";
+import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run";
 import type { Authenticator } from "@app/lib/auth";
 import { AgentDustAppRunConfiguration } from "@app/lib/models/assistant/actions/dust_app_run";
 import { AppResource } from "@app/lib/resources/app_resource";
