+++ front/lib/models/assistant/actions/dust_app_run.ts
@@ -1,7 +1,7 @@
 import type { CreationOptional, ForeignKey, NonAttribute } from "sequelize";
 import { DataTypes } from "sequelize";
 
-import type { DustAppParameters } from "@app/lib/actions/types/dust_app_run";
+import type { DustAppParameters } from "@app/lib/actions/dust_app_run";
 import { AgentConfiguration } from "@app/lib/models/assistant/agent";
 import { AgentMessage } from "@app/lib/models/assistant/conversation";
 import { frontSequelize } from "@app/lib/resources/storage";
